+++ modules/src/events.rs
@@ -36,6 +36,7 @@ pub enum IbcEvent {
 
     CreateClient(ClientEvents::CreateClient),
     UpdateClient(ClientEvents::UpdateClient),
+    UpgradeClient(ClientEvents::UpgradeClient),
     ClientMisbehaviour(ClientEvents::ClientMisbehaviour),
 
     OpenInitConnection(ConnectionEvents::OpenInit),
@@ -113,6 +114,7 @@ impl IbcEvent {
             IbcEvent::NewBlock(ev) => ev.set_height(height),
             IbcEvent::CreateClient(ev) => ev.set_height(height),
             IbcEvent::UpdateClient(ev) => ev.set_height(height),
+            IbcEvent::UpgradeClient(ev) => ev.set_height(height),
             IbcEvent::ClientMisbehaviour(ev) => ev.set_height(height),
             IbcEvent::OpenInitConnection(ev) => ev.set_height(height),
             IbcEvent::OpenTryConnection(ev) => ev.set_height(height),