+++ modules/src/ics02_client/events.rs
@@ -17,6 +17,7 @@ use crate::ics24_host::identifier::ClientId;
 const CREATE_EVENT_TYPE: &str = "create_client";
 const UPDATE_EVENT_TYPE: &str = "update_client";
 const MISBEHAVIOUR_EVENT_TYPE: &str = "client_misbehaviour";
+const UPGRADE_EVENT_TYPE: &str = "upgrade_client";
 
 /// The content of the `key` field for the attribute containing the client identifier.
 const CLIENT_ID_ATTRIBUTE_KEY: &str = "client_id";
@@ -40,6 +41,9 @@ pub fn try_from_tx(event: &tendermint::abci::Event) -> Option<IbcEvent> {
             header: extract_header_from_tx(event), // TODO fix
         })),
         MISBEHAVIOUR_EVENT_TYPE => Some(IbcEvent::ClientMisbehaviour(ClientMisbehaviour(
+            extract_attributes_from_tx(event),
+        ))),
+        UPGRADE_EVENT_TYPE => Some(IbcEvent::UpgradeClient(UpgradeClient(
             extract_attributes_from_tx(event),
         ))),
         _ => None,
@@ -102,7 +106,7 @@ impl From<NewBlock> for IbcEvent {
     }
 }
 
-#[derive(Debug, Deserialize, Serialize, Clone)]
+#[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct Attributes {
     pub height: Height,
     pub client_id: ClientId,
@@ -256,5 +260,21 @@ impl TryFrom<RawObject> for ClientMisbehaviour {
 impl From<ClientMisbehaviour> for IbcEvent {
     fn from(v: ClientMisbehaviour) -> Self {
         IbcEvent::ClientMisbehaviour(v)
+    }
+}
+
+/// Signals a recent upgrade of an on-chain client (IBC Client).
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Deserialize, Serialize)]
+pub struct UpgradeClient(Attributes);
+
+impl UpgradeClient {
+    pub fn set_height(&mut self, height: Height) {
+        self.0.height = height;
+    }
+}
+
+impl From<Attributes> for UpgradeClient {
+    fn from(attrs: Attributes) -> Self {
+        UpgradeClient(attrs)
     }
 }