+++ modules/src/ics02_client/msgs.rs
@@ -7,10 +7,12 @@
 use crate::ics02_client::msgs::create_client::MsgCreateAnyClient;
 use crate::ics02_client::msgs::misbehavior::MsgSubmitAnyMisbehaviour;
 use crate::ics02_client::msgs::update_client::MsgUpdateAnyClient;
+use crate::ics02_client::msgs::upgrade_client::MsgUpgradeAnyClient;
 
 pub mod create_client;
 pub mod misbehavior;
 pub mod update_client;
+pub mod upgrade_client;
 
 #[allow(clippy::large_enum_variant)]
 #[derive(Clone, Debug)]
@@ -18,4 +20,5 @@ pub enum ClientMsg {
     CreateClient(MsgCreateAnyClient),
     UpdateClient(MsgUpdateAnyClient),
     Misbehaviour(MsgSubmitAnyMisbehaviour),
+    UpgradeClient(MsgUpgradeAnyClient),
 }