+++ modules/src/ics04_channel/error.rs
@@ -37,6 +37,9 @@ pub enum Kind {
     #[error("invalid proof: missing height")]
     MissingHeight,
 
+    #[error("Missing sequence number for receiving packets")]
+    MissingNextRecvSeq,
+
     #[error("packet sequence cannot be 0")]
     ZeroPacketSequence,
 
@@ -54,6 +57,9 @@ pub enum Kind {
 
     #[error("there is no packet in this message")]
     MissingPacket,
+
+    #[error("Packet with the sequence number {0} has been already received")]
+    PacketAlreadyReceived(Sequence),
 
     #[error("missing counterparty")]
     MissingCounterparty,
@@ -95,6 +101,15 @@ pub enum Kind {
     #[error("Channel chain verification fails on ChannelOpenTry for ChannelOpenInit")]
     FailedChanneOpenTryVerification,
 
+    #[error("Verification fails for the packet with the sequence number {0}")]
+    PacketVerificationFailed(Sequence),
+
+    #[error("Acknowledgment cannot be empty")]
+    InvalidAcknowledgement,
+
+    #[error("Packet acknowledgement exists for the packet with the sequence {0}")]
+    AcknowledgementExists(Sequence),
+
     #[error("No client state associated with client id {0}")]
     MissingClientState(ClientId),
 
@@ -134,8 +149,17 @@ pub enum Kind {
     #[error("Handshake proof verification fails at ChannelOpenAck")]
     ChanOpenAckProofVerification,
 
+    #[error("Commitment for the packet {0} not found")]
+    PacketCommitmentNotFound(Sequence),
+
     #[error("Handshake proof verification fails at ChannelOpenConfirm")]
     ChanOpenConfirmProofVerification,
+
+    #[error("The stored commitment of the packet {0} is incorrect")]
+    IncorrectPacketCommitment(Sequence),
+
+    #[error("Missing sequence number for ack packets")]
+    MissingNextAckSeq,
 }
 
 impl Kind {