+++ proto/src/lib.rs
@@ -10,6 +10,7 @@
 
 /// The version (commit hash) of the Cosmos SDK used when generating this library.
 pub const COSMOS_SDK_VERSION: &str = include_str!("prost/COSMOS_SDK_COMMIT");
+pub const COSMOS_IBC_VERSION: &str = include_str!("prost/COSMOS_IBC_COMMIT");
 
 pub mod cosmos {
     pub mod auth {
@@ -69,13 +70,23 @@ pub mod cosmos {
             include!("prost/cosmos.tx.v1beta1.rs");
         }
     }
+    pub mod upgrade {
+        pub mod v1beta1 {
+            include!("prost/cosmos.upgrade.v1beta1.rs");
+        }
+    }
+    pub mod gov {
+        pub mod v1beta1 {
+            include!("prost/cosmos.gov.v1beta1.rs");
+        }
+    }
 }
 
 pub mod ibc {
-    pub mod applications {
+    pub mod apps {
         pub mod transfer {
             pub mod v1 {
-                include!("prost/ibc.applications.transfer.v1.rs");
+                include!("prost/ibc.apps.transfer.v1.rs");
             }
         }
     }