+++ proto/src/prost/ibc.lightclients.localhost.v1.rs
@@ -4,8 +4,8 @@
 pub struct ClientState {
     /// self chain ID
     #[prost(string, tag="1")]
-    pub chain_id: std::string::String,
+    pub chain_id: ::prost::alloc::string::String,
     /// self latest block height
     #[prost(message, optional, tag="2")]
-    pub height: ::std::option::Option<super::super::super::core::client::v1::Height>,
+    pub height: ::core::option::Option<super::super::super::core::client::v1::Height>,
 }