+++ relayer-cli/src/commands/create.rs
@@ -1,7 +1,9 @@
 //! `create` subcommand
 use abscissa_core::{Command, Help, Options, Runnable};
 
+use crate::commands::create::channel::CreateChannelCommand;
 use crate::commands::create::connection::CreateConnectionCommand;
+use crate::commands::tx::client::TxCreateClientCmd;
 
 mod channel;
 mod connection;
@@ -13,10 +15,15 @@ pub enum CreateCmds {
     #[options(help = "Get usage information")]
     Help(Help<Self>),
 
+    /// Subcommand for creating a `client`
+    #[options(help = "Create a new IBC client")]
+    Client(TxCreateClientCmd),
+
     /// Subcommand for creating a `connection`
     #[options(help = "Create a new connection between two chains")]
     Connection(CreateConnectionCommand),
-    // /// Subcommand for creating a `channel` with various
-    // #[options(help = "create a new channel between two chains")]
-    // Channel(CreateChannelCommand),
+
+    /// Subcommand for creating a `channel`
+    #[options(help = "Create a new channel between two chains")]
+    Channel(CreateChannelCommand),
 }