+++ relayer/src/chain/cosmos.rs
@@ -39,8 +39,8 @@ use ibc::ics23_commitment::merkle::convert_tm_to_ics_merkle_proof;
 use ibc::ics24_host::identifier::{ChainId, ChannelId, ClientId, ConnectionId, PortId};
 use ibc::ics24_host::Path::ClientConsensusState as ClientConsensusPath;
 use ibc::ics24_host::Path::ClientState as ClientStatePath;
-use ibc::ics24_host::{Path, IBC_QUERY_PATH};
 use ibc::query::QueryTxRequest;
+use ibc::ics24_host::{ClientUpgradePath, Path, IBC_QUERY_PATH, SDK_UPGRADE_QUERY_PATH};
 use ibc::signer::Signer;
 use ibc::Height as ICSHeight;
 // Support for GRPC
@@ -48,6 +48,9 @@ use ibc_proto::cosmos::auth::v1beta1::{BaseAccount, QueryAccountRequest};
 use ibc_proto::cosmos::base::v1beta1::Coin;
 use ibc_proto::cosmos::tx::v1beta1::mode_info::{Single, Sum};
 use ibc_proto::cosmos::tx::v1beta1::{AuthInfo, Fee, ModeInfo, SignDoc, SignerInfo, TxBody, TxRaw};
+use ibc_proto::cosmos::upgrade::v1beta1::{
+    QueryCurrentPlanRequest, QueryUpgradedConsensusStateRequest,
+};
 use ibc_proto::ibc::core::channel::v1::{
     PacketState, QueryChannelsRequest, QueryConnectionChannelsRequest,
     QueryNextSequenceReceiveRequest, QueryPacketAcknowledgementsRequest,
@@ -73,6 +76,7 @@ use super::Chain;
 const DEFAULT_MAX_GAS: u64 = 300000;
 const DEFAULT_MAX_MSG_NUM: usize = 30;
 const DEFAULT_MAX_TX_SIZE: usize = 2 * 1048576; // 2 MBytes
+const DEFAULT_GAS_FEE_AMOUNT: u64 = 1000;
 
 pub struct CosmosSdkChain {
     config: ChainConfig,
@@ -162,6 +166,8 @@ impl CosmosSdkChain {
         let mut pk_buf = Vec::new();
         prost::Message::encode(&key.public_key.public_key.to_bytes(), &mut pk_buf).unwrap();
 
+        crate::time!("PK {:?}", hex::encode(key.public_key.public_key.to_bytes()));
+
         // Create a MsgSend proto Any message
         let pk_any = Any {
             type_url: "/cosmos.crypto.secp256k1.PubKey".to_string(),
@@ -181,14 +187,8 @@ impl CosmosSdkChain {
             sequence: acct_response.sequence,
         };
 
-        // Gas Fee
-        let coin = Coin {
-            denom: "stake".to_string(),
-            amount: "1000".to_string(),
-        };
-
         let fee = Some(Fee {
-            amount: vec![coin],
+            amount: vec![self.fee()],
             gas_limit: self.gas(),
             payer: "".to_string(),
             granter: "".to_string(),
@@ -207,7 +207,7 @@ impl CosmosSdkChain {
             body_bytes: body_buf.clone(),
             auth_info_bytes: auth_buf.clone(),
             chain_id: self.config.clone().id.to_string(),
-            account_number: 0,
+            account_number: acct_response.account_number,
         };
 
         // A protobuf serialization of a SignDoc
@@ -226,6 +226,8 @@ impl CosmosSdkChain {
         let mut txraw_buf = Vec::new();
         prost::Message::encode(&tx_raw, &mut txraw_buf).unwrap();
 
+        crate::time!("TxRAW {:?}", hex::encode(txraw_buf.clone()));
+
         let response = self
             .block_on(broadcast_tx_commit(self, txraw_buf))
             .map_err(|e| Kind::Rpc(self.config.rpc_addr.clone()).context(e))?;
@@ -237,6 +239,19 @@ impl CosmosSdkChain {
 
     fn gas(&self) -> u64 {
         self.config.gas.unwrap_or(DEFAULT_MAX_GAS)
+    }
+
+    fn fee(&self) -> Coin {
+        let amount = self
+            .config
+            .clone()
+            .fee_amount
+            .unwrap_or(DEFAULT_GAS_FEE_AMOUNT);
+
+        Coin {
+            denom: self.config.fee_denom.clone(),
+            amount: amount.to_string(),
+        }
     }
 
     fn max_msg_num(&self) -> usize {
@@ -267,6 +282,34 @@ impl CosmosSdkChain {
         if prove {}
 
         Ok(response)
+    }
+
+    // Perform an ABCI query against the client upgrade sub-store to fetch a proof.
+    fn query_client_upgrade_proof(
+        &self,
+        data: ClientUpgradePath,
+        height: Height,
+    ) -> Result<(MerkleProof, ICSHeight), Error> {
+        let prev_height =
+            Height::try_from(height.value() - 1).map_err(|e| Kind::InvalidHeight.context(e))?;
+
+        let path = TendermintABCIPath::from_str(SDK_UPGRADE_QUERY_PATH).unwrap();
+        let response = self.block_on(abci_query(
+            &self,
+            path,
+            Path::Upgrade(data).to_string(),
+            prev_height,
+            true,
+        ))?;
+
+        let proof = response.proof.ok_or(Kind::EmptyResponseProof)?;
+
+        let height = ICSHeight::new(
+            self.config.id.version(),
+            response.height.increment().value(),
+        );
+
+        Ok((proof, height))
     }
 }
 
@@ -476,6 +519,105 @@ impl Chain for CosmosSdkChain {
         Ok(client_state)
     }
 
+    fn query_upgraded_client_state(
+        &self,
+        height: ICSHeight,
+    ) -> Result<(Self::ClientState, MerkleProof), Error> {
+        crate::time!("query_upgraded_client_state");
+
+        let grpc_address =
+            Uri::from_str(&self.config.grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
+
+        let mut client = self
+            .block_on(
+                ibc_proto::cosmos::upgrade::v1beta1::query_client::QueryClient::connect(
+                    grpc_address,
+                ),
+            )
+            .map_err(|e| Kind::Grpc.context(e))?;
+
+        let req = tonic::Request::new(QueryCurrentPlanRequest {});
+        let response = self
+            .block_on(client.current_plan(req))
+            .map_err(|e| Kind::Grpc.context(e))?;
+
+        let upgraded_client_state_raw = response
+            .into_inner()
+            .plan
+            .ok_or(Kind::EmptyResponseValue)?
+            .upgraded_client_state
+            .ok_or(Kind::EmptyUpgradedClientState)?;
+        let client_state = AnyClientState::try_from(upgraded_client_state_raw)
+            .map_err(|e| Kind::Grpc.context(e))?;
+
+        // TODO: Better error kinds here.
+        let tm_client_state =
+            downcast!(client_state => AnyClientState::Tendermint).ok_or_else(|| {
+                Kind::Query("upgraded client state".into()).context("unexpected client state type")
+            })?;
+
+        // Query for the proof.
+        let tm_height =
+            Height::try_from(height.revision_height).map_err(|e| Kind::InvalidHeight.context(e))?;
+        let (proof, _proof_height) = self.query_client_upgrade_proof(
+            ClientUpgradePath::UpgradedClientState(height.revision_height),
+            tm_height,
+        )?;
+
+        Ok((tm_client_state, proof))
+    }
+
+    fn query_upgraded_consensus_state(
+        &self,
+        height: ICSHeight,
+    ) -> Result<(Self::ConsensusState, MerkleProof), Error> {
+        crate::time!("query_upgraded_consensus_state");
+
+        let tm_height =
+            Height::try_from(height.revision_height).map_err(|e| Kind::InvalidHeight.context(e))?;
+
+        let grpc_address =
+            Uri::from_str(&self.config.grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
+
+        let mut client = self
+            .block_on(
+                ibc_proto::cosmos::upgrade::v1beta1::query_client::QueryClient::connect(
+                    grpc_address,
+                ),
+            )
+            .map_err(|e| Kind::Grpc.context(e))?;
+
+        let req = tonic::Request::new(QueryUpgradedConsensusStateRequest {
+            last_height: tm_height.into(),
+        });
+        let response = self
+            .block_on(client.upgraded_consensus_state(req))
+            .map_err(|e| Kind::Grpc.context(e))?;
+
+        let upgraded_consensus_state_raw = response
+            .into_inner()
+            .upgraded_consensus_state
+            .ok_or(Kind::EmptyResponseValue)?;
+
+        // TODO: More explicit error kinds (should not reuse Grpc all over the place)
+        let consensus_state = AnyConsensusState::try_from(upgraded_consensus_state_raw)
+            .map_err(|e| Kind::Grpc.context(e))?;
+
+        let tm_consensus_state = downcast!(consensus_state => AnyConsensusState::Tendermint)
+            .ok_or_else(|| {
+                Kind::Query("upgraded consensus state".into())
+                    .context("unexpected consensus state type")
+            })?;
+
+        // Fetch the proof.
+        let (proof, _proof_height) = self.query_client_upgrade_proof(
+            ClientUpgradePath::UpgradedClientConsensusState(height.revision_height),
+            tm_height,
+        )?;
+
+        Ok((tm_consensus_state, proof))
+    }
+
     /// Performs a query to retrieve the identifiers of all connections.
     fn query_consensus_states(
         &self,
@@ -579,7 +721,7 @@ impl Chain for CosmosSdkChain {
     ) -> Result<ConnectionEnd, Error> {
         let res = self.query(Path::Connections(connection_id.clone()), height, false)?;
         Ok(ConnectionEnd::decode_vec(&res.value)
-            .map_err(|e| Kind::Query("connection".into()).context(e))?)
+            .map_err(|e| Kind::Query(format!("connection {}", connection_id)).context(e))?)
     }
 
     fn query_connection_channels(
@@ -1043,13 +1185,11 @@ impl Chain for CosmosSdkChain {
 
     fn build_header(
         &self,
+        trusted_height: ICSHeight,
         trusted_light_block: Self::LightBlock,
         target_light_block: Self::LightBlock,
     ) -> Result<Self::Header, Error> {
         crate::time!("build_header");
-
-        let trusted_height =
-            ICSHeight::new(self.id().version(), trusted_light_block.height().into());
 
         Ok(TmHeader {
             trusted_height,