+++ relayer/src/chain/handle.rs
@@ -130,7 +130,17 @@ pub enum ChainRequest {
 
     QueryConsensusStates {
         request: QueryConsensusStatesRequest,
-        reply_to: ReplyTo<Vec<AnyConsensusStateWithHeight>>,
+        reply_to: ReplyTo<Vec<AnyConsensusStateWithHeight> >,
+    },
+
+    QueryUpgradedClientState {
+        height: Height,
+        reply_to: ReplyTo<(AnyClientState, MerkleProof)>,
+    },
+
+    QueryUpgradedConsensusState {
+        height: Height,
+        reply_to: ReplyTo<(AnyConsensusState, MerkleProof)>,
     },
 
     QueryCommitmentPrefix {
@@ -253,6 +263,16 @@ pub trait ChainHandle: DynClone + Send + Sync + Debug {
         &self,
         request: QueryConsensusStatesRequest,
     ) -> Result<Vec<AnyConsensusStateWithHeight>, Error>;
+
+    fn query_upgraded_client_state(
+        &self,
+        height: Height,
+    ) -> Result<(AnyClientState, MerkleProof), Error>;
+
+    fn query_upgraded_consensus_state(
+        &self,
+        height: Height,
+    ) -> Result<(AnyConsensusState, MerkleProof), Error>;
 
     fn query_commitment_prefix(&self) -> Result<CommitmentPrefix, Error>;