+++ relayer/src/foreign_client.rs
@@ -2,7 +2,7 @@ use std::{thread, time::Duration};
 
 use prost_types::Any;
 use thiserror::Error;
-use tracing::{error, info};
+use tracing::{debug, error, info, warn};
 
 use ibc::downcast;
 use ibc::events::{IbcEvent, IbcEventType};
@@ -14,6 +14,7 @@ use ibc::ics02_client::header::Header;
 use ibc::ics02_client::msgs::create_client::MsgCreateAnyClient;
 use ibc::ics02_client::msgs::misbehavior::MsgSubmitAnyMisbehaviour;
 use ibc::ics02_client::msgs::update_client::MsgUpdateAnyClient;
+use ibc::ics02_client::msgs::upgrade_client::MsgUpgradeAnyClient;
 use ibc::ics24_host::identifier::{ChainId, ClientId};
 use ibc::query::QueryTxRequest;
 use ibc::tx_msg::Msg;
@@ -38,6 +39,9 @@ pub enum ForeignClientError {
 
     #[error("error raised while submitting the misbehaviour evidence: {0}")]
     Misbehaviour(String),
+
+    #[error("failed while trying to upgrade client id {0} with error: {1}")]
+    ClientUpgrade(ClientId, String),
 }
 
 #[derive(Clone, Debug)]
@@ -128,6 +132,87 @@ impl ForeignClient {
                 format!("{}", e),
             )),
         }
+    }
+
+    pub fn upgrade(&self) -> Result<Vec<IbcEvent>, ForeignClientError> {
+        // Fetch the latest height of the source chain.
+        let src_height = self.src_chain.query_latest_height().map_err(|e| {
+            ForeignClientError::ClientUpgrade(
+                self.id.clone(),
+                format!(
+                    "failed while querying src chain ({}) for latest height: {}",
+                    self.src_chain.id(),
+                    e
+                ),
+            )
+        })?;
+
+        info!("Upgrade Height: {}", src_height);
+
+        let mut msgs = self.build_update_client(src_height)?;
+
+        // Query the host chain for the upgraded client state, consensus state & their proofs.
+        let (client_state, proof_upgrade_client) = self
+            .src_chain
+            .query_upgraded_client_state(src_height)
+            .map_err(|e| {
+                ForeignClientError::ClientUpgrade(
+                    self.id.clone(),
+                    format!(
+                        "failed while fetching from chain {} the upgraded client state: {}",
+                        self.src_chain.id(),
+                        e
+                    ),
+                )
+            })?;
+
+        debug!("Upgraded client state {:?}", client_state);
+
+        let (consensus_state, proof_upgrade_consensus_state) = self
+            .src_chain
+            .query_upgraded_consensus_state(src_height)
+            .map_err(|e| ForeignClientError::ClientUpgrade(self.id.clone(), format!(
+                "failed while fetching from chain {} the upgraded client consensus state: {}", self.src_chain.id(), e)))
+            ?;
+
+        debug!("Upgraded client consensus state {:?}", consensus_state);
+
+        // Get signer
+        let signer = self.dst_chain.get_signer().map_err(|e| {
+            ForeignClientError::ClientUpgrade(
+                self.id.clone(),
+                format!(
+                    "failed while fetching the destination chain ({}) signer: {}",
+                    self.dst_chain.id(),
+                    e
+                ),
+            )
+        })?;
+
+        let msg_upgrade = MsgUpgradeAnyClient {
+            client_id: self.id.clone(),
+            client_state,
+            consensus_state,
+            proof_upgrade_client,
+            proof_upgrade_consensus_state,
+            signer,
+        }
+        .to_any();
+
+        msgs.push(msg_upgrade);
+
+        let res = self.dst_chain.send_msgs(msgs).map_err(|e| {
+            ForeignClientError::ClientUpgrade(
+                self.id.clone(),
+                format!(
+                    "failed while sending message to destination chain {} with err: {}",
+                    self.dst_chain.id(),
+                    e
+                ),
+            )
+        })?;
+
+        Ok(res)
     }
 
     /// Returns a handle to the chain hosting this client.
@@ -231,6 +316,8 @@ impl ForeignClient {
         Ok(())
     }
 
+    /// Returns a vector with a message for updating the client to height `target_height`.
+    /// If the client already stores consensus states for this height, returns an empty vector.
     pub fn build_update_client(
         &self,
         target_height: Height,
@@ -257,6 +344,14 @@ impl ForeignClient {
                 ))
             })?
             .latest_height();
+
+        if trusted_height >= target_height {
+            warn!(
+                "Client height ({}) >= chain target height ({}). Cannot build update message.",
+                trusted_height, target_height
+            );
+            return Ok(vec![]);
+        }
 
         let header = self
             .src_chain()
@@ -293,7 +388,16 @@ impl ForeignClient {
                 e
             ))
         })?;
+
         let new_msgs = self.build_update_client(h)?;
+        if new_msgs.is_empty() {
+            return Err(ForeignClientError::ClientUpdate(format!(
+                "Client {} is already up-to-date with chain {}@{}",
+                self.id,
+                self.src_chain.id(),
+                h
+            )));
+        }
 
         let mut events = self.dst_chain().send_msgs(new_msgs).map_err(|e| {
             ForeignClientError::ClientUpdate(format!(